Moved functions with `Scalar` as `lhs` into separate `ScalarTests` file
